requirements.txt
Flask==2.3.3
pytube==15.0.0
openai-whisper==20231117